+++ app/src/main/java/com/ar/arstoken/data/db/ItemSalesRow.kt
@@ -3,6 +3,6 @@ package com.ar.arstoken.data.db
 data class ItemSalesRow(
     val itemId: Int,
     val itemName: String,
-    val totalQty: Double,
-    val totalAmount: Double
+    val totalQty: Int,
+    val totalAmount: Int
 )
